Moved calendar related functions from user.services to a new calendar.services file
